# openapi.yaml
openapi: 3.1.0
info:
  title: Artic Hub
  version: 0.1.0-alpha
paths:
...
  /api/market/price/{symbol}:
    get:
      tags:
      - market
      summary: Price One
      operationId: price_one_api_market_price__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/market/candles:
    get:
      tags:
      - market
      summary: Candles
      operationId: candles_api_market_candles_get
      parameters:
      - name: symbol
        in: query
        required: true
        schema:
          type: string
          title: Symbol
      - name: interval
        in: query
        required: false
        schema:
          type: string
          default: 15m
          title: Interval
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError